+++ lib/src/internal/editor.dart
@@ -9,14 +9,24 @@ import '../types/entity.dart';
 import 'plugin.dart';
 
 class Editor {
-  final IosEditor _iOS = const IosEditor();
+  final IosEditor _ios = const IosEditor();
+  final DarwinEditor _darwin = const DarwinEditor();
   final AndroidEditor _android = const AndroidEditor();
 
+  @Deprecated('Use `Editor.darwin` instead. This will be removed in 3.0.0')
   IosEditor get iOS {
-    if (Platform.isIOS) {
-      return _iOS;
+    if (Platform.isIOS || Platform.isMacOS) {
+      return _ios;
     }
     throw const OSError('iOS Editor should only be use on iOS.');
+  }
+
+  /// Support iOS and macOS.
+  DarwinEditor get darwin {
+    if (Platform.isIOS || Platform.isMacOS) {
+      return _darwin;
+    }
+    throw const OSError('Darwin Editor should only be use on iOS or macOS.');
   }
 
   AndroidEditor get android {
@@ -118,8 +128,8 @@ class Editor {
   }
 }
 
-class IosEditor {
-  const IosEditor();
+class DarwinEditor {
+  const DarwinEditor();
 
   /// {@template photo_manager.IosEditor.EnsureParentIsRootOrFolder}
   /// Folders and albums can be only created under the root path or folders,
@@ -210,6 +220,33 @@ class IosEditor {
     }
     return null;
   }
+
+  /// Save Live Photo to the gallery from the given [imageFile] and [videoFile].
+  ///
+  /// {@macro photo_manager.Editor.TitleWhenSaving}
+  ///
+  /// {@macro photo_manager.Editor.DescriptionWhenSaving}
+  ///
+  /// {@macro photo_manager.Editor.SavingAssets}
+  Future<AssetEntity?> saveLivePhoto({
+    required File imageFile,
+    required File videoFile,
+    required String title,
+    String? desc,
+    String? relativePath,
+  }) {
+    return plugin.saveLivePhoto(
+      imageFile: imageFile,
+      videoFile: videoFile,
+      title: title,
+      desc: desc,
+      relativePath: relativePath,
+    );
+  }
+}
+
+class IosEditor extends DarwinEditor {
+  const IosEditor();
 }
 
 class AndroidEditor {