+++ CHANGELOG.md
@@ -4,6 +4,48 @@ that can be found in the LICENSE file. -->
 
 # CHANGELOG
 
+## 2.5.2
+
+### Improvements
+
+- Use `ContentUris` for retrieving Media URIs on Android. (#870)
+
+## 2.5.1
+
+### Improvements
+
+- Improve media subtype on iOS. (#863)
+
+## 2.5.0
+
+### Features
+
+- Support saving Live Photos on iOS and macOS. (#851)
+- Introduce `DarwinEditor` to replace `IosEditor`. (#855)
+
+## 2.4.2
+
+### Improvements
+
+- Expose `frame` for `AssetEntity.thumbnailDataWithSize`. (#850)
+
+## 2.4.1
+
+### Improvements
+
+- Use last modified date for Glide caches key on Android. (#848)
+
+## 2.4.0
+
+### Features
+
+- Support both legacy and scoped storage on Android. (#833)
+
+### Fixes
+
+- Avoid duplicate `copyItemAtURL` for videos on iOS. (#840)
+- Correct permission checks with `requestPermissionExtend` on Android 13. (#843)
+
 ## 2.3.0
 
 ### Features
@@ -456,7 +498,7 @@ and applied multiple ### Fixes which make this plugin as the most solid ever.
   but at the same time user have to bear the risks corresponding to the permission).
 - Support clean file cache.
 - Experimental
-    - Preload image (Use `PhotoCachingManager` api.)
+  - Preload image (Use `PhotoCachingManager` api.)
 - Add `OrderOption` as sort condition. The option default value is order by create date desc;
 - Support icloud asset progress.
 
@@ -519,14 +561,14 @@ and applied multiple ### Fixes which make this plugin as the most solid ever.
 - Create AssetEntity with id.
 - Create AssetPathEntity from id.
 - Only iOS
-    - Create folder or album.
-    - Remove assets in album.
-    - Delete folder or album.
-    - Favorite asset.
+  - Create folder or album.
+  - Remove assets in album.
+  - Delete folder or album.
+  - Favorite asset.
 - Only android
-    - move asset to another path.
-    - Remove all non-existing rows.
-    - add `relativePath` for android.
+  - move asset to another path.
+  - Remove all non-existing rows.
+  - add `relativePath` for android.
 
 ### Improvements
 
@@ -548,6 +590,7 @@ and applied multiple ### Fixes which make this plugin as the most solid ever.
 ## 0.5.0
 
 ### Breaking changes
+
 - Add date condition to filter datetime
 - Add class `DateTimeCond`
 - Add `dateTimeCond` to `FilterOptionGroup`